+++ src/parser/enrichers/spell/UnseenServant.mjs
@@ -1,0 +1,21 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class UnseenServant extends DDBEnricherData {
+
+  get type() {
+    return "summon";
+  }
+
+  get activity() {
+    return {
+      noTemplate: true,
+      profileKeys: [
+        "UnseenServantSRD",
+      ],
+      summons: {
+      },
+    };
+  }
+
+}